daug kas

+++ resources/views/profilis.blade.php
@@ -21,13 +21,14 @@
                   <div class="d-flex flex-column align-items-center text-center">
                     <img src="/uploads/avatars/{{$data->avatar}}" alt="Image" class="rounded-circle" width="150">
                     <div class="mt-3">
-                      <span class="badge badge-secondary">Atsijunges</span>
+                    
                       <h4>{{$data->name}}</h4>
                       @if( $data->lygis > 0)
                         <div class="RankBadge Rank{{$data->lygis}}"></div>
                       @endif
                       <p class="text-secondary mb-1">Atlikti užsakymai</p><b>{{$data->uzsakymuKiekis}}</b>
                       <p class="text-muted font-size-sm">{{$data->miestas}}</p>
+                      @if (\Illuminate\Support\Facades\Auth::check())
                         @if(App\Http\Controllers\IsimintiController::arVartotojasIsimintas($data->id, Auth::user()->id) == false && $data->id != Auth::user()->id)
                       <form action="/isimintiVartotojaPrideti">
                         <input type="hidden" name="vartotojoid" id="vartotojoid" value="{{ $data->id }}">
@@ -35,6 +36,7 @@
                           <button type="submit" class="btn btn-primary">Įsiminti</button>
                         </form>
                         @endif
+                      @endif
                           <a class="btn btn-outline-primary" href="/zinutes/{{$data->id}}">Rašyti</a>
                     </div>
                   </div>
@@ -109,10 +111,12 @@
                   <hr>
                   <div class="row">
                     <div class="col-sm-12">
+                      @if (\Illuminate\Support\Facades\Auth::check())
                         @if ($data->id ==  Auth::user()->id)
                          <a class="btn btn-info " href="/profilisEdit">Nustatymai</a>
                          <a class="btn btn-info " href="/keistiSlaptazodi">Keisti slaptažodį</a>
                         @endif
+                      @endif
                     </div>
                   </div>
                 </div>
@@ -129,6 +133,7 @@
                                 <div class="d-flex flex-row align-items-center align-content-center post-title"><span class="bdge mr-1">Viso</span><span class="mr-2 comments"> {{App\Http\Controllers\ProfilisController::countAtsiliepimus( $data->id )}} atsiliepimai-(ų)&nbsp;</span></div>
                             </div>
                         </div>
+                        @if (\Illuminate\Support\Facades\Auth::check())
                         <form class="form-style-5"role="form" method="POST" action="/submitAtsiliepima"> 
                         @csrf  
                           <div class="coment-bottom bg-white p-2 px-4">
@@ -138,18 +143,21 @@
                               <button class="btn btn-primary" type="submitAtsiliepima">Pateikti</button>
                           </div>
                         </form>
+                        @endif
                           @foreach($atsiliepimai as $atsiliepimas)
                             <div class="commented-section mt-2">
                                 <div class="d-flex flex-row align-items-center commented-user">
                                     <h5 class="mr-2">{{$atsiliepimas->userKomentavo->name}}</h5><span class="dot mb-1"></span><span class="mb-1 ml-2">{{$atsiliepimas->data}}</span>
-                                    @if ($atsiliepimas->kas_komentavo == Auth::user()->id)
-                                    <form class="form-style-5"role="form" method="GET" action="/deleteAtsiliepima/{{$atsiliepimas->id}}"> 
-                                    @csrf  
-                                    <button type="submit" class="btn-close" onclick="javascript:return confirm('Ar tikrai norite pašalinti atsiliepimą?')" href='deleteAtsiliepima/{{$atsiliepimas->id}}'>
-                                    <span class="icon-cross"></span>
-                                    <span class="visually-hidden">Close</span>
-                                    </button>
-                                    </form>
+                                    @if (\Illuminate\Support\Facades\Auth::check())
+                                      @if ($atsiliepimas->kas_komentavo == Auth::user()->id)
+                                      <form class="form-style-5"role="form" method="GET" action="/deleteAtsiliepima/{{$atsiliepimas->id}}"> 
+                                      @csrf  
+                                      <button type="submit" class="btn-close" onclick="javascript:return confirm('Ar tikrai norite pašalinti atsiliepimą?')" href='deleteAtsiliepima/{{$atsiliepimas->id}}'>
+                                      <span class="icon-cross"></span>
+                                      <span class="visually-hidden">Close</span>
+                                      </button>
+                                      </form>
+                                      @endif
                                     @endif
                                 </div>
                                 <div class="comment-text-sm"><span>{{$atsiliepimas->tekstas}}.</span></div>
